scripts/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
jsonschema==4.19.2
PyYAML==6.0.1